move LocalCache into InstanceCache
Benchmarks show that this is more efficient since the ExecContext in the interpreter's hot path can be more lightweight.
